+++ src/cuda/gpuCamera.cu
@@ -8,18 +8,18 @@ using namespace CMU462;
 class gpuCamera {
   public:
     
-    __host__ gpuCamera(Matrix3x3 c2w, Vector3D p, size_t screenW, size_t screenH, double screenDist);
+    __host__ gpuCamera(Matrix3x3 c2w, Vector3D p, size_t screenW, size_t screenH, float screenDist);
     
-    __device__ gpuRay generate_ray(double x, double y);
+    __device__ gpuRay generate_ray(float x, float y);
 
     gpuVector3D pos;
     gpuMatrix3x3 c2w;
     size_t screenW, screenH;
-    double screenDist;
+    float screenDist;
 
 };
 gpuCamera::gpuCamera(Matrix3x3 cameraToWorld, Vector3D p, 
-    size_t screenW, size_t screenH, double screenDist) 
+    size_t screenW, size_t screenH, float screenDist) 
   : screenW(screenW), screenH(screenH), screenDist(screenDist)
 {
   pos = gpuVector3D(p.x, p.y, p.z);
@@ -32,7 +32,7 @@ gpuCamera::gpuCamera(Matrix3x3 cameraToWorld, Vector3D p,
   }
 }
 
-__device__ gpuRay gpuCamera::generate_ray(double x, double y) {
+__device__ gpuRay gpuCamera::generate_ray(float x, float y) {
   gpuVector3D sp_cam = gpuVector3D(-(x - 0.5) * screenW / screenDist,
                                    -(y - 0.5) * screenH / screenDist, 1);
   gpuVector3D dir_cam = -sp_cam;
